Moved configuration to provider object
Using a value for configuration is useful for simple values,
but not for objects that needs to be extended instead of overwritten.
The usual way to do this is by creating a provider that can be
used to configure the module, so that's what has been done here.

Fly-by:
Update with the latest generator version